+++ sync-for-commerce/src/codatsynccommerce/models/operations/get_sync_flow_url.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
+from ..shared import errormessage as shared_errormessage
 from ..shared import syncflowurl as shared_syncflowurl
 from typing import Optional
 
@@ -25,6 +26,8 @@ class GetSyncFlowURLResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     sync_flow_url: Optional[shared_syncflowurl.SyncFlowURL] = dataclasses.field(default=None)