+++ sync-for-commerce/src/codatsynccommerce/models/operations/get_sync_status.py
@@ -3,8 +3,9 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
+from ..shared import errormessage as shared_errormessage
 from ..shared import syncstatus as shared_syncstatus
-from typing import Any, Optional
+from typing import Optional
 
 
 @dataclasses.dataclass
@@ -21,10 +22,8 @@ class GetSyncStatusResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    bad_request: Optional[Any] = dataclasses.field(default=None)
-    r"""Bad Request"""
-    not_found: Optional[Any] = dataclasses.field(default=None)
-    r"""Not Found"""
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     sync_status: Optional[shared_syncstatus.SyncStatus] = dataclasses.field(default=None)