+++ sync-for-commerce/src/codatsynccommerce/models/operations/request_sync_for_date_range.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
+from ..shared import errormessage as shared_errormessage
 from ..shared import syncrange as shared_syncrange
 from ..shared import syncsummary as shared_syncsummary
 from typing import Optional
@@ -23,6 +24,8 @@ class RequestSyncForDateRangeResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     sync_summary: Optional[shared_syncsummary.SyncSummary] = dataclasses.field(default=None)