+++ sync-for-commerce/RELEASES.md
@@ -559,3 +559,13 @@ Based on:
 - [python v3.0.0] sync-for-commerce
 ### Releases
 - [PyPI v3.0.0] https://pypi.org/project/codat-sync-for-commerce/3.0.0 - sync-for-commerce
+
+## 2023-10-30 01:25:46
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0] sync-for-commerce
+### Releases
+- [PyPI v3.1.0] https://pypi.org/project/codat-sync-for-commerce/3.1.0 - sync-for-commerce